% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ams_iih_ontology.R
\name{read_ams_iih_facility_shapes}
\alias{read_ams_iih_facility_shapes}
\title{Load AMS-II.H facility type applicability shape}
\usage{
read_ams_iih_facility_shapes()
}
\value{
A \code{sh_shape_graph} as returned by \code{\link[shapeR:read_shacl]{shapeR::read_shacl()}}.
}
\description{
Load AMS-II.H facility type applicability shape
}
